refactor(ui): extract MainWindow and modularize application bootstrap in main.py
